update comments after some testing and debugging

+++ process.h
@@ -26,7 +26,7 @@ struct ProcessControl
 #define STACK_SIZE PAGE_SIZE
 #define TOTAL_PROCS 10
 #define USERSPACE_CONTEXT_SIZE (12*8) /* 12 GPRs saved on the stack when context switch done by scheduler (see swap function) */
-#define REGISTER_POSITION(addr, n) ((addr) + (n*8)) /* Position of nth 8-byte register from current address */
+#define REGISTER_POSITION(addr, n) ((uint64_t)(addr) + (n*8)) /* Position of nth 8-byte register from current address */
 
 enum En_ProcessState
 {
